refactor(templates): consolidate message templates into a dedicated file for improved organization
- Moved `twitterMessageHandlerTemplate`, `twitterShouldRespondTemplate`, `twitterSearchTemplate`, `twitterSpaceFillerTemplate`, and `twitterSpaceTopicSuggestionTemplate` into `templates.ts` for better modularity and maintainability.
- Updated imports in `interactions.ts`, `search.ts`, and `spaces.ts` to utilize the new template structure, promoting cleaner code practices and adherence to clean code principles.
